# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AIS Catering",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
        "/caterings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering"
                ],
                "summary": "Returns list of caterings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "used for pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "used for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of caterings",
                        "schema": {
                            "$ref": "#/definitions/GetCateringsResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering"
                ],
                "summary": "Returns error or 201 status code if success",
                "parameters": [
                    {
                        "description": "Catering Name",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/AddNameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "catering object",
                        "schema": {
                            "$ref": "#/definitions/CateringsResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/caterings/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering"
                ],
                "summary": "Returns info about catering",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "catering model",
                        "schema": {
                            "$ref": "#/definitions/CateringsResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering"
                ],
                "summary": "Returns 204 if success and 4xx error if failed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Catering Name",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/AddNameRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully updated"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering"
                ],
                "summary": "Soft delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/caterings/{id}/schedules": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "caterings schedules"
                ],
                "summary": "Returns list of schedules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CateringSchedule"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/caterings/{id}/schedules/{scheduleId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "caterings schedules"
                ],
                "summary": "Returns 200 and updated model if success and 4xx error if failed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CateringSchedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CateringSchedule model",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CateringSchedule model",
                        "schema": {
                            "$ref": "#/definitions/CateringSchedule"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AddNameRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "aisnovations"
                }
            }
        },
...
        "CateringSchedule": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isWorking": {
                    "type": "boolean"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "CateringsResponse": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "GetCateringsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CateringsResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "UpdateScheduleRequest": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "isWorking": {
                    "type": "boolean"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
    }
}